Update processes_b.c
updated the critical section, there's a structure with student id and mark that stores the student number from the student list, mark component is there for easy reference and might take it down later. semaphore lock for the first TA and the next TA so two processes can access the data. once accessed take the student id then unlock the semaphore. mark the student and store it.